+++ lib/trailblazer/activity/dsl/linear.rb
@@ -28,21 +28,6 @@ class Trailblazer::Activity
             options # {id: "Start.success"}
           ]
         end
-
-        class IndexError < IndexError
-          attr_reader :step_id
-
-          def initialize(sequence, step_id)
-            @step_id  = step_id
-            valid_ids = sequence.collect{ |row| row[3][:id].inspect }
-
-            message = "\n" \
-              "\e[31m#{@step_id.inspect} is not a valid step ID. Did you mean any of these ?\e[0m\n" \
-              "\e[32m#{valid_ids.join("\n")}\e[0m"
-
-            super(message)
-          end
-        end
       end
 
       # Sequence