+++ CHANGES.md
@@ -30,6 +30,7 @@
 * Make `Strategy` a class. It makes constant management much simpler to understand.
 * Move `Insert` to `trailblazer-activity` gem.
 * We now need `Sequence::Row` instances in `Sequence` to adhere to the Adds specification.
+* Move `Sequence::IndexError` to `Activity::Adds::IndexError` in the `trailblazer-activity` gem. Remove `IndexError#step_id`.
 
 # 0.5.0